- bug 1639: write/truncate lock inversion
 - bug 1763: fix timestamps from jumping to "now"
 - bug 1648: extra journal assertions
 - bug 1771: add an extra multiunlink test
 - bug 1776: fix read_record/write_record API
 - bug 1772: fix leak of offset_extent, possible incorrect i_size later
 - bug 1541: fix lasserts in mis-matched transnos during open-unlink testing
 - don't mark a file as an orphan unless this is the last link
 - avoid committing NULL handle in force close
 - local.sh is now a one-stripe LOV configuration

+++ lustre/tests/small_write.c
@@ -72,7 +72,8 @@ int main (int argc, char **argv) {
 	rc = fstat(fd, &st);
 	if (rc < 0 || st.st_size != bytes) {
 		printf("bad file %lu size first write %lu != %lu: rc %d\n",
-		       st.st_ino, st.st_size, bytes, rc);
+		       (unsigned long)st.st_ino, (unsigned long)st.st_size,
+                       bytes, rc);
 		return 1;
 	}
 
@@ -91,7 +92,8 @@ int main (int argc, char **argv) {
 	rc = fstat(fd, &st);
 	if (rc < 0 || st.st_size != bytes + bytes / 2) {
 		printf("bad file %lu size second write %lu != %lu: rc %d\n",
-		       st.st_ino, st.st_size, bytes, rc);
+		       (unsigned long)st.st_ino, (unsigned long)st.st_size,
+                       bytes, rc);
 		return 1;
 	}
 
@@ -111,7 +113,8 @@ int main (int argc, char **argv) {
 		rc = fstat(fd, &st);
 		if (rc < 0 || st.st_size != bytes + bytes / 2) {
 			printf("bad file size after read %lu != %lu: rc %d\n",
-			       st.st_size, bytes + bytes / 2, rc);
+			       (unsigned long)st.st_size, bytes + bytes / 2,
+                               rc);
 			return 1;
 		}
